API bug fixes
